+++ src/loc_gallery/__main__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from avv_gallery.server import run
+from loc_gallery.server import run
 
 if __name__ == "__main__":
     run()